Add Basic Home Study Functionality
User is prompted to begin their session. An alarm triggers at intervals specified in the list of alarms (in minutes).
Trigger times and responses are captured and presented in a list. Alarm make noise when triggered and either
times out or is dismissed when the control is used to respond to the alarm.

The intents are intense.